+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - added ability to list projects in cloud storage services for model loading
 - server evaluation endpoint at ``POST /evaluate``
 - CRF entity recognizer now returns a confidence score when extracting entities
+- server endpoint at ``DELETE /models`` to unload models from server memory
 
 Changed
 -------